move proposals related files
